Notification service added.

+++ src/app/pages/cars/cars.component.ts
@@ -3,6 +3,7 @@ import { ToastrService } from 'ngx-toastr';
 import { Car } from 'src/app/models/car';
 import { CarDetailDto } from 'src/app/models/DTOs/carDetailDto';
 import { CarService } from 'src/app/services/car.service';
+import { NotificationService } from 'src/app/services/notification.service';
 
 @Component({
   selector: 'app-cars',
@@ -13,7 +14,9 @@ export class CarsComponent implements OnInit {
   dataLoaded = false;
   carDetails: CarDetailDto[] = [];
 
-  constructor(private carService: CarService, private toastrService:ToastrService) { }
+  constructor(private carService: CarService, 
+    private toastrService: ToastrService, 
+    private notificationService: NotificationService) { }
 
   ngOnInit(): void {
     this.getCarDetails();
@@ -27,15 +30,23 @@ export class CarsComponent implements OnInit {
     })
   }
 
-  editCar(carDetail:CarDetailDto){
+  addCar() {
+    this.notificationService.showSuccess("Added.");
+  }
+
+  editCar(carDetail: CarDetailDto) {
     console.log("edit", carDetail);
   }
 
-  removeCar(carDetail:CarDetailDto){
-    console.log("remove", carDetail);
-    this.carService.deleteCarWithCarDetailDto(carDetail).subscribe(rp => {
-      this.toastrService.info("Succesfully deleted!");
-      console.log("Deleted!", rp);
-    })
+  removeCar(carDetail: CarDetailDto) {
+    if (confirm("Are you use you want to delete this car: " + carDetail.id + " / " + carDetail.brand.name + " / " + carDetail.name)) {
+      this.carService.deleteCarWithCarDetailDto(carDetail).subscribe(rp => {
+        this.toastrService.info("Succesfully deleted!");
+        console.log("Deleted!", rp);
+      })
+    }
+    else {
+      this.toastrService.info("Car not deleted.");
+    }
   }
 }
